+++ apps/codeforafrica/src/components/FooterLinks/index.js
@@ -1,0 +1,3 @@
+import FooterLinks from "./FooterLinks";
+
+export default FooterLinks;
